% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_selection.R
\name{bic_selection}
\alias{bic_selection}
\title{BIC method for selection of hyper-parameters of phi and rho}
\usage{
bic_selection(
  Y,
  q,
  V,
  n_subject,
  preprocess = TRUE,
  penalty = "L1",
  phi_grid_search = seq(1, 2, 0.2),
  rho_grid_search = c(0.9),
  maxIteration = 100,
  espli1 = 0.01,
  espli2 = 0.05,
  demean = TRUE,
  save_output = FALSE
)
}
\arguments{
\item{Y}{Group-level dynamic connectivity data is represented as a matrix of dimension \code{NT}*\code{p},
where \code{NT} denotes the number of subjects multiplied by the number of time points,
and \code{p} represents the number of edges in the connectivity network.}

\item{q}{The number of connectivity traits to extract.}

\item{V}{The number of nodes. Note that \code{p} (the columns of \code{Y}) needs to be equal to \code{(V-1)V/2}.}

\item{n_subject}{The total number of subjects in the study.}

\item{preprocess}{If \code{TRUE}, the concatenated group-level connectivity data will be preprocessed. Defaults to \code{TRUE}.}

\item{penalty}{The option for the penalization function for the sparsity regularization for the connectivity traits.
The users can choose "NULL"","L1", or "SCAD". Defaults to "L1".}

\item{phi_grid_search}{Grid search candidates of tuning parameter \code{phi} for penalty on connectivity traits. Defaults to \code{seq(1, 2, 0.2)}.}

\item{rho_grid_search}{Grid search candidates of tuning parameter for the adaptive selection method to determine the number of ranks
for modeling each connectivity trait. Defaults to \code{c(0.9)}.}

\item{maxIteration}{The maximum number of iterations. Defaults to 100.}

\item{espli1}{A number describing the tolerance for change on \code{A}.}

\item{espli2}{A number describing the tolerance for change on \code{S}.}

\item{demean}{If \code{TRUE}, performs demeaning on the input data. Defaults to \code{TRUE}.}

\item{save_output}{If \code{TRUE}, saves the output. Defaults to \code{FALSE}.}
}
\value{
A list containing the following elements:
\item{bic_tab}{A dataframe containing BIC values for each combination of \code{phi} and \code{rho}.}
\item{results}{A list of dyna-LOCUS outputs, if \code{save_output} is set to \code{TRUE}.}
}
\description{
BIC method for selection of hyper-parameters of phi and rho
}
