% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_bic.R
\name{calculate_bic}
\alias{calculate_bic}
\title{Calculate the Bayesian Information Criterion (BIC) for a given model.}
\usage{
calculate_bic(Y, A, S)
}
\arguments{
\item{Y}{The observed data matrix.}

\item{A}{The loading matrix.}

\item{S}{The latent source matrix.}
}
\value{
The BIC value for the model.
}
\description{
This function computes the BIC for a dynamic model represented by the matrices A and S.
}
