% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccf_calculation.R
\name{ccf_calculation}
\alias{ccf_calculation}
\title{Calculate cross-correlation function (CCF) for connectivity traits}
\usage{
ccf_calculation(A, q, n_subject)
}
\arguments{
\item{A}{The trait loading matrix.}

\item{q}{The number of connectivity traits.}

\item{n_subject}{The number of subjects.}
}
\value{
An array containing the final CCF values for each pair of connectivity traits.
}
\description{
This function calculates the cross-correlation function (CCF) between pairs of connectivity traits.
}
\details{
The function follows these steps:
\itemize{
\item For each subject, calculate the lag which achieves the largest absolute CCF between trait pairs.
\item Determine the sign of the CCF for each trait pair.
\item Recalculate the maximum value lag between connectivity trait pairs using the determined sign.
\item Determine the mode of lag for each connectivity trait pair.
\item Calculate and return the final CCF values.
}
}
