% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dFC_states.R
\name{dFC_states}
\alias{dFC_states}
\title{Retrieve Whole Brain dFC States Based on dyna-LOCUS Results}
\usage{
dFC_states(A, S, n_subject, q, n_examplers = 10, gap = 5, seed = 1, n_centers)
}
\arguments{
\item{A}{The loading matrix.}

\item{S}{The dyna-LOCUS result connectivity traits.}

\item{n_subject}{The total number of subjects in the study.}

\item{q}{The number of connectivity traits.}

\item{n_examplers}{The number of exemplars to select from each subject for clustering initialization. Default is 10.}

\item{gap}{The gap between selected exemplars in time points. Default is 5.}

\item{seed}{An integer for setting the random seed to ensure reproducibility. Default is 1.}

\item{n_centers}{The number of centers (or clusters) to use in k-means clustering.}
}
\value{
A matrix of reconstructed connectivity patterns based on the clustered loading matrix. Each row of the matrix represents the upper triangular part of a dFC state.
}
\description{
This function retrieves whole-brain dynamic functional connectivity (dFC) states based on the results from the dyna-LOCUS method.
}
