% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyna_initial.R
\name{dyna_initial}
\alias{dyna_initial}
\title{Generate starting values for the iteration}
\usage{
dyna_initial(Y, q, V, rho = 0.9, R = NULL, maxIter = 100)
}
\arguments{
\item{Y}{A demeaned concatenated group-level connectivity data after whitening if preprocess = TRUE}

\item{q}{The number of latent sources one would like to extract}

\item{V}{The number of nodes in the study}

\item{rho}{A proportional tuning parameter ranging from 0 to 1 for the adaptive selection method to determine the number of ranks
for modeling each connectivity trait. The value of rho represents the closeness of the connectivity traits estimated with and
without the low-rank structure. A higher value of rho will lead to a higher rank. Defaults to 0.9.}

\item{R}{A numeric vector of pre-specifying ranks of latent sources. Defaults to NULL.}

\item{maxIter}{The maximum iteration number when doing ica}
}
\value{
Initial values of loading matrix \code{A} and latent source \code{S}
}
\description{
Generate starting values for the iteration
}
