% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyna_preprocess.R
\name{dyna_preprocess}
\alias{dyna_preprocess}
\title{Preprocess the connectivity data through whitening}
\usage{
dyna_preprocess(Ynew, q)
}
\arguments{
\item{Ynew}{A demeaned concatenated group-level connectivity data before whitening}

\item{q}{The number of latent sources one would like to extract}
}
\value{
whitened Y (Y tilde) and the dewhitening matrix H_star
}
\description{
Preprocess the connectivity data through whitening
}
