% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyna_update.R
\name{dyna_update}
\alias{dyna_update}
\title{Iterative Update Algorithm for dyna-LOCUS}
\usage{
dyna_update(
  Y,
  A,
  S,
  theta,
  q,
  K,
  V,
  n_subject,
  t_length,
  penalty_S,
  phi,
  penalty_A,
  lambda,
  H_star,
  list_sparse,
  preprocess,
  speed_up,
  silent
)
}
\arguments{
\item{Y}{The (preprocessed) concatenated group-level connectivity data.}

\item{A}{A loading matrix from the previous iteration.}

\item{theta}{A list containing trait-specific parameters and channels.}

\item{q}{The number of latent connectivity traits to extract.}

\item{K}{The number of edges.}

\item{V}{The number of nodes.}

\item{n_subject}{The total number of subjects in the study.}

\item{t_length}{The number of sliding windows.}

\item{penalty_S}{The option for the penalization function for the sparsity regularization for the connectivity traits.
Users can choose "NULL", "L1", or "SCAD". Defaults to "L1".}

\item{phi}{A tuning parameter for the element-wise penalty on S. Defaults to 2.}

\item{penalty_A}{The option for the temporal smoothness regularization for the trait loadings.
If \code{TRUE}, a temporal smoothness penalty is included in the optimization function to encourage similarity in the trait loadings
in adjacent time windows. Defaults to \code{TRUE}.}

\item{lambda}{A tuning parameter for the temporal smooth penalty on \code{A}.}

\item{H_star}{The dewhitening matrix.}

\item{list_sparse}{A list to record the sparseness of connectivity traits.}

\item{preprocess}{If \code{TRUE}, the concatenated group-level connectivity data is preprocessed.}

\item{speed_up}{If \code{TRUE}, the updating process is sped up using singular value decomposition.}

\item{silent}{If \code{TRUE}, suppresses the printout of penalties applied on \code{A} and \code{S}.}
}
\description{
Iterative Update Algorithm for dyna-LOCUS
}
