% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynaLOCUS.R
\name{dynaLOCUS}
\alias{dynaLOCUS}
\title{A Novel Decomposition Method for Brain Network Dynamic Connectivity Matrices
Using Low-Rank Structure with Uniform Sparsity and Temporal Smoothness Regularization}
\usage{
dynaLOCUS(
  Y,
  q,
  V,
  n_subject,
  preprocess = TRUE,
  penalty_S = "L1",
  phi = 2,
  rho = 0.9,
  penalty_A = TRUE,
  lambda = 0.1,
  maxIteration = 100,
  speed_up = FALSE,
  espli1 = 0.01,
  espli2 = 0.05,
  demean = TRUE,
  silent = TRUE
)
}
\arguments{
\item{Y}{Group-level dynamic connectivity data, represented as a matrix of dimension \code{NT} \times \code{p},
where \code{NT} denotes the number of subjects multiplied by the number of time points,
and \code{p} represents the number of edges in the connectivity network.}

\item{q}{The number of connectivity traits to extract.}

\item{V}{The number of nodes. Note that \code{p} (the columns of \code{Y}) needs to be equal to \code{(V-1)V/2}.}

\item{n_subject}{The total number of subjects in the study.}

\item{preprocess}{If \code{TRUE}, the concatenated group-level connectivity data will be preprocessed. Defaults to \code{TRUE}.}

\item{penalty_S}{The option for the penalization function for the sparsity regularization for the connectivity traits.
Users can choose "NULL", "L1", or "SCAD". Defaults to "L1".}

\item{phi}{A tuning parameter for the element-wise penalty on \code{S}. Defaults to 2.}

\item{rho}{A proportional tuning parameter ranging from 0 to 1 for the adaptive selection method to determine the number of ranks
for modeling each connectivity trait. The value of \code{rho} represents the closeness of the connectivity traits estimated with and
without the low-rank structure. A higher value of \code{rho} will lead to a higher rank. Defaults to 0.9.}

\item{penalty_A}{The option for the temporal smoothness regularization for the trait loadings.
If \code{TRUE}, a temporal smoothness penalty is included in the optimization function to encourage similarity in the trait loadings
in adjacent time windows. Defaults to \code{TRUE}.}

\item{lambda}{A numeric tuning parameter for the temporal smooth lasso penalty on \code{A}. Defaults to 0.1.}

\item{maxIteration}{The maximum number of iterations. Defaults to 100.}

\item{speed_up}{If \code{FALSE}, use the Node-rotation algorithm for learning dyna-LOCUS. If \code{TRUE}, use the Alternative algorithm to
further speed up the computation for learning dyna-LOCUS. Defaults to \code{FALSE}.}

\item{espli1}{A number describing the tolerance for change on \code{A}.}

\item{espli2}{A number describing the tolerance for change on \code{S}.}

\item{demean}{If \code{TRUE}, demean each column of the group-level connectivity matrix \code{Y}. Defaults to \code{TRUE}.}

\item{silent}{If \code{TRUE}, suppress the output. Defaults to \code{TRUE}.}
}
\value{
The resulting loading matrix \code{A} and connectivity traits \code{S} based on the decomposition method.
}
\description{
A Novel Decomposition Method for Brain Network Dynamic Connectivity Matrices
Using Low-Rank Structure with Uniform Sparsity and Temporal Smoothness Regularization
}
