% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_var_cal.R
\name{energy_var_cal}
\alias{energy_var_cal}
\title{Calculate energy and variation for connectivity traits}
\usage{
energy_var_cal(A, q, t_length)
}
\arguments{
\item{A}{The loading matrix.}

\item{q}{The number of connectivity traits.}

\item{t_length}{The length of the time series per subject.}
}
\value{
A list containing:
\item{trait_energy}{A vector representing the mean log-transformed energy of each connectivity trait across subjects.}
\item{trait_variation}{A vector representing the mean variation of each connectivity trait across subjects.}
}
\description{
This function calculates the energy and variation for connectivity traits.
}
