% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_selection.R
\name{lambda_selection}
\alias{lambda_selection}
\title{Goodness-of-Fit Method for Hyper-Parameter Lambda Selection}
\usage{
lambda_selection(
  Y,
  q,
  V,
  n_subject,
  preprocess = TRUE,
  penalty_S = "L1",
  phi = 2,
  rho = 0.9,
  lambda_grid_search = c(0.001, 0.005, 0.01, 0.05, 0.1),
  maxIteration = 100,
  espli1 = 0.01,
  espli2 = 0.05,
  demean = TRUE,
  save_output = FALSE
)
}
\arguments{
\item{Y}{Group-level dynamic connectivity data represented as a matrix of dimension \code{NT \times p},
where \code{NT} denotes the number of subjects multiplied by the number of time points,
and \code{p} represents the number of edges in the connectivity network.}

\item{q}{The number of connectivity traits to extract.}

\item{V}{The number of nodes. Note that \code{p} (the columns of \code{Y}) needs to be equal to \code{(V-1)V/2}.}

\item{n_subject}{The total number of subjects in the study.}

\item{preprocess}{If \code{TRUE}, the concatenated group-level connectivity data will be preprocessed. Defaults to \code{TRUE}.}

\item{penalty_S}{The option for the penalization function for the sparsity regularization for the connectivity traits.
Users can choose "NULL", "L1", or "SCAD". Defaults to "L1".}

\item{phi}{A tuning parameter for the element-wise penalty on \code{S}. Defaults to 2.}

\item{rho}{A proportional tuning parameter ranging from 0 to 1 for the adaptive selection method to determine the number of ranks
for modeling each connectivity trait. The value of \code{rho} represents the closeness of the connectivity traits estimated with and
without the low-rank structure. A higher value of \code{rho} will lead to a higher rank. Defaults to 0.9.}

\item{lambda_grid_search}{A vector of candidate values for the tuning parameter \code{lambda}.
Defaults to \code{c(NA, 1e-3, 5e-3, 1e-2, 5e-2, 1e-1)}, where \code{NA} represents no temporal smoothness penalty.}

\item{maxIteration}{The maximum number of iterations. Defaults to 100.}

\item{espli1}{A number describing the tolerance for change on \code{A}.}

\item{espli2}{A number describing the tolerance for change on \code{S}.}

\item{demean}{If \code{TRUE}, demean each column of the group-level connectivity matrix \code{Y}. Defaults to \code{TRUE}.}

\item{save_output}{If \code{TRUE}, saves the output. Defaults to \code{FALSE}.}
}
\value{
A list containing the following elements:
\item{selected_lambda}{The lambda value that achieves the minimal goodness-of-fit.}
\item{goodness_of_fit}{Goodness-of-fit values for each \code{lambda}.}
\item{results}{A list of dyna-LOCUS outputs, if \code{save_output} is set to \code{TRUE}.}
}
\description{
Goodness-of-Fit Method for Hyper-Parameter Lambda Selection
}
