% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Ltrans}
\alias{Ltrans}
\title{Helper function to transform the upper triangular of a matrix to a vector}
\usage{
Ltrans(X, d = F)
}
\arguments{
\item{X}{A symmetric matrix}

\item{d}{If \code{True}, include the diagonal part of the matrix X. Defaults to \code{FALSE}.}
}
\description{
Helper function to transform the upper triangular of a matrix to a vector
}
