% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Ltrinv}
\alias{Ltrinv}
\title{Helper function to transform a vector to a matrix}
\usage{
Ltrinv(x, V, d = F)
}
\arguments{
\item{V}{The dimension of the target matrix}

\item{d}{If \code{TRUE}, the vector X include the diagonal part of the matrix. Defaults to \code{FALSE}.}

\item{X}{A vector representing the upper triangular of a matrix}
}
\description{
Helper function to transform a vector to a matrix
}
