% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matsplitter}
\alias{matsplitter}
\title{Split a matrix into submatrices}
\usage{
matsplitter(M, r, c)
}
\arguments{
\item{M}{A matrix to be split.}

\item{r}{The number of rows in each submatrix.}

\item{c}{The number of columns in each submatrix.}
}
\value{
An array where each slice along the third dimension represents a submatrix of dimensions \code{r} by \code{c}.
}
\description{
This function splits a given matrix \code{M} into smaller submatrices of dimensions \code{r} by \code{c}.
}
