% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability_index.R
\name{reliability_index}
\alias{reliability_index}
\title{Reliability Index Calculation}
\usage{
reliability_index(
  Y,
  q_choices,
  V,
  n_subject,
  preprocess = TRUE,
  penalty_S = "L1",
  phi = 2,
  rho = 0.9,
  penalty_A = TRUE,
  lambda = 0.1,
  maxIteration = 100,
  espli1 = 0.01,
  espli2 = 0.05,
  silent = TRUE,
  demean = TRUE,
  seeds = 1:50
)
}
\arguments{
\item{Y}{Group-level dynamic connectivity data, represented as a matrix of dimension \code{NT} \times \code{p},
where \code{NT} denotes the number of subjects multiplied by the number of time points,
and \code{p} represents the number of edges in the connectivity network.}

\item{q_choices}{A vector of integers representing different choices for the number of connectivity traits to extract.}

\item{V}{The number of nodes. Note that \code{p} (the columns of \code{Y}) needs to be equal to \code{(V-1)V/2}.}

\item{n_subject}{The total number of subjects in the study.}

\item{preprocess}{If \code{TRUE}, the concatenated group-level connectivity data will be preprocessed. Defaults to \code{TRUE}.}

\item{penalty_S}{The option for the penalization function for the sparsity regularization for the connectivity traits.
Users can choose "NULL", "L1", or "SCAD". Defaults to "L1".}

\item{phi}{A tuning parameter for the element-wise penalty on S. Defaults to 2.}

\item{rho}{A proportional tuning parameter ranging from 0 to 1 for the adaptive selection method to determine the number of ranks
for modeling each connectivity trait. The value of \code{rho} represents the closeness of the connectivity traits estimated with and
without the low-rank structure. A higher value of \code{rho} will lead to a higher rank. Defaults to 0.9.}

\item{penalty_A}{The option for the temporal smoothness regularization for the trait loadings.
If \code{TRUE}, a temporal smoothness penalty is included in the optimization function to encourage similarity in the trait loadings
in adjacent time windows. Defaults to \code{TRUE}.}

\item{lambda}{A numeric tuning parameter for the temporal smooth lasso penalty on \code{A}. Defaults to 0.1.}

\item{maxIteration}{The maximum number of iterations for the dyna-LOCUS algorithm.}

\item{espli1}{A number describing the tolerance for change on A.}

\item{espli2}{A number describing the tolerance for change on S.}

\item{silent}{Logical, if FALSE, print out the penalty added on A and S.}

\item{demean}{Logical, if TRUE, demean the data before processing.}

\item{seeds}{A vector of integers to set seeds for bootstrap sampling, default is 1:50.}
}
\value{
A list containing:
\item{reliability}{A list where each component corresponds to a different choice of q. Each component is of length q,
representing the reliability index for each trait based on that choice of q.}
\item{RI}{A list where each component corresponds to a different choice of q. Each component is of length q,
representing the reproducibility (RI) for each trait based on that choice of q.}
}
\description{
This function calculates the reliability index and reproducibility (RI) for the dyna-LOCUS method.
}
