% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SCAD_func}
\alias{SCAD_func}
\title{Helper function to calculate the scad penalty}
\usage{
SCAD_func(yvpen, phi = 2, gamma = 2.1)
}
\arguments{
\item{yvpen}{A vector upon penalization}

\item{phi}{A tunning parameter}

\item{gamma}{A tunning parameter}
}
\description{
Helper function to calculate the scad penalty
}
